+++ uefi/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Implemented `PartialEq<char>` for `Char8` and `Char16`.
 - Added `CStr16::from_char16_with_nul` and `Char16::from_char16_with_nul_unchecked`.
 - Added terminal GUID constants to `device_path::messaging::Vendor`.
+- Added `MemoryMap::from_raw`.
 
 ## Changed
 - `DevicePath::to_string` and `DevicePathNode::to_string` now return